funcionando metodos USER, continuo con recetas
